requirements.txt
PyYAML
sh
zuul
